Removed Firebase
Though definitely worthwhile for a large amount of users, it was much faster and more efficient to have all of my data stored locally and filtered on the page itself.
I've achieved this with Tabulator. Though I did attempt to use DataTables, attempting to filter using that was a nightmare, whereas in Tabulator it's a simple "setFilter" or "addFilter" etc., very simple! Love it.